refactor cleanup - prep for skellycam
