+++ skellysnapshot/backend/pose_estimation_2d/mediapipe_things/mediapipe_config.py
