+++ skellysnapshot/gui/widgets/skeleton_view_widget.py
@@ -1,7 +1,8 @@
 
-from PyQt6.QtWidgets import QWidget,QVBoxLayout
 
 import matplotlib
+from PySide6.QtWidgets import QWidget, QVBoxLayout
+
 matplotlib.use('QtAgg')
 from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg
 from matplotlib.figure import Figure
@@ -11,8 +12,8 @@ import numpy as np
 import matplotlib.pyplot as plt
 import numpy as np
 
-from skellysnapshot.reconstruction_3d.snapshot_3d_dataclass import SnapshotData3d
-from skellysnapshot.visualize_3d.mediapipe_bone_connections import build_mediapipe_skeleton
+from skellysnapshot.backend.reconstruction_3d.snapshot_3d_dataclass import SnapshotData3d
+from skellysnapshot.backend.visualize_3d.mediapipe_bone_connections import build_mediapipe_skeleton
 
 
 class SkeletonViewWidget(QWidget):
